refactored the widgets folder
