+++ lib/app/widgets/cards/activity.card.dart
