+++ lib/app/widgets/cards/video.card.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-import '../core/core.dart';
+import '../../core/core.dart';
 
 class VideoCard extends StatelessWidget {
   final String image;
