+++ lib/app/widgets/messages/text.suggestion.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-import '../core/core.dart';
+import '../../core/core.dart';
 
 class TextSuggestions extends StatelessWidget {
   const TextSuggestions({
