+++ lib/app/widgets/other/side.stream.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-import '../core/core.dart';
+import '../../core/core.dart';
 
 class SideStream extends StatelessWidget {
   const SideStream({
