+++ lib/app/widgets/other/stars.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-import '../core/core.dart';
+import '../../core/core.dart';
 
 class StarsWidget extends StatelessWidget {
   final double sizeH;
