+++ lib/app/widgets/widgets.dart
@@ -1,16 +1,16 @@
-export 'activity.card.dart';
-export 'app.bar.dart';
-export 'app.button.dart';
-export 'app.radio.dart';
-export 'bottom.message.box.dart';
-export 'custom.listtile.dart';
-export 'custom.switch.dart';
-export 'dance.gridview.card.dart';
-export 'gridview.card.dart';
-export 'icon.card.dart';
-export 'scroll.card.dart';
-export 'side.stream.dart';
-export 'stars.dart';
-export 'stream.tile.dart';
-export 'text.suggestion.dart';
-export 'video.card.dart';
+export 'cards/activity.card.dart';
+export 'other/app.bar.dart';
+export 'other/app.button.dart';
+export 'other/app.radio.dart';
+export 'messages/bottom.message.box.dart';
+export 'tiles/custom.listtile.dart';
+export 'other/custom.switch.dart';
+export 'cards/dance.gridview.card.dart';
+export 'cards/gridview.card.dart';
+export 'cards/icon.card.dart';
+export 'cards/scroll.card.dart';
+export 'other/side.stream.dart';
+export 'other/stars.dart';
+export 'tiles/stream.tile.dart';
+export 'messages/text.suggestion.dart';
+export 'cards/video.card.dart';
